# Add comprehensive frontend API testing with fail-fast integration architecture