--- a/Dockerfile.prod
+++ b/Dockerfile.prod
@@ -110,7 +110,7 @@ HEALTHCHECK --interval=30s --timeout=10s --start-period=5s --retries=3 \
     CMD ["./starter", "health-check"] || exit 1
 
 # Expose port
-EXPOSE 8080
+EXPOSE 3000
 
 # Default command
-CMD ["./starter", "server", "--port", "8080"]
+CMD ["./starter", "server", "--port", "3000"]