--- a/web/e2e/api-health.spec.ts
+++ b/web/e2e/api-health.spec.ts
@@ -34,6 +34,7 @@ test.describe('API Health Checks', () => {
   });
 
   test('OpenAPI spec is accessible', async ({ request }) => {
+    // Proxied to backend API through vite dev server
     const response = await request.get('/api-docs/openapi.json');
     await expect(response).toBeOK();