--- a/web/vitest.config.integration.ts
+++ b/web/vitest.config.integration.ts
@@ -1,0 +1,18 @@
+import { defineConfig } from 'vitest/config';
+import { resolve } from 'path';
+
+export default defineConfig({
+  test: {
+    environment: 'jsdom',
+    setupFiles: ['./src/test/setup.ts'],
+    include: ['src/**/*.integration.test.ts'],
+    globals: true,
+    testTimeout: 30000,
+    pool: 'forks',
+  },
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, './src'),
+    },
+  },
+});