… CLAUDE.md
… README.md
… docker-compose.chaos.yaml
… docker-compose.prod.yaml
… docs/ADVANCED-TOPICS.md
… docs/API-REFERENCE.md
… docs/ARCHITECTURE-GUIDE.md
… docs/IMPLEMENTATION-GUIDE.md
… docs/PRODUCTION-DEPLOYMENT.md
… docs/QUICK-START.md
… docs/README.md
… docs/TROUBLESHOOTING.md
… docs/ideas/ai-first.md
… docs/openapi.json
… k8s/deployment.yaml
… scripts/CHAOS_GUIDE.md
… web/README.md
… web/package.json
… web/registration-debug.png
… web/src/hooks/__tests__/useApiQueries.test.ts
… web/src/lib/__tests__/cross-feature.integration.test.ts
… web/src/lib/__tests__/rbac-integration.integration.test.ts
… web/src/lib/__tests__/user-journeys.integration.test.ts
… web/src/lib/api/__tests__/client.integration.test.ts
… web/src/lib/api/__tests__/client.test.ts
… web/src/lib/auth/__tests__/guards.test.ts
… web/src/lib/rbac/__tests__/types.test.ts
… web/src/test/README.md
… web/src/test/integration-setup.ts
… web/src/test/mocks.ts