@quanhua92 2025-08-10
Thank you @gemini-code-assist for the excellent and thorough code review\! 🙏

I've addressed all 4 medium-priority suggestions you identified:

## ✅ **Fixes Applied:**

### 1. **CLAUDE.md - Trailing Space Cleanup**
- Removed trailing space from `pnpm test:integration` command reference
- Improved documentation hygiene as suggested

### 2. **E2E Auth Test - Deterministic Button Wait**
- **Before**: `await page.waitForTimeout(500)` ❌ (flaky)
- **After**: `await expect(button).toBeEnabled()` ✅ (deterministic)
- Eliminates race conditions and makes tests more robust

### 3. **Package.json - Maintainable Glob Patterns** 
- **Before**: Hardcoded file paths requiring manual updates
- **After**: `src/**/*.integration.test.ts` glob pattern with dedicated `vitest.config.integration.ts`
- New integration test files now automatically included - much more maintainable\!

### 4. **Error Recovery Test - Proper Pass/Fail Conditions**
- **Before**: `expect(true).toBe(true)` always passed on timeout ❌
- **After**: Proper error throwing with descriptive failure message ✅  
- Test now correctly fails when error recovery mechanism is too slow

## 🧪 **Testing Results:**
- ✅ E2E auth flow passes in 4.7s with deterministic button check
- ✅ Integration test glob pattern correctly finds both test files  
- ✅ Error recovery test now has proper failure semantics
- ✅ All functionality maintained while improving code quality

Your feedback was spot-on for improving test robustness and maintainability. The codebase is now even more reliable and follows industry best practices\!

Thanks for helping make this frontend testing infrastructure production-ready\! 🚀

---
*All changes committed in [7c75544](https://github.com/quanhua92/rust-fullstack-starter/commit/7c75544)*